requirements.txt
openai
python-dotenv
gunicorn
flask
wikipedia-api